GeneticAlgorithms.Utils.print_sudoku — Method

print_sudoku(chromosome::Chromosome)

Prints the Sudoku genes of the given chromosome.

Arguments

chromosome::Chromosome: The chromosome to print.

Example

GeneticAlgorithms.Utils.visualize_function_with_contours — Method

Plots a contour of a 2D function over the specified x and y ranges.

Arguments

f::Function: The function to visualize.

x_range: Range of x-values (default: (-2.0, 2.0)).

y_range: Range of y-values (default: (-2.0, 2.0)).

path: Optimization path points (default: []).

Returns

A contour plot of the function with optional highlighted points.

GeneticAlgorithms.Utils.visualize_rosenbrock_results — Method

visualize_rosenbrock_results(best_chromosomes::Vector{Chromosome}, save_path::String="result.png")

Generates a visualization of the optimization process for the rosenbrock fitness function with given best chromosomes.

Arguments

best_chromosomes: A sequence of optimization results (chromosomes) for visualizing the optimization path.

Example

julia
visualize_results([Chromosome([0.0, 0.0]), Chromosome([0.5, 0.5]), Chromosome([1.0, 1.0])], "rosenbrock_path.png")