GeneticAlgorithms

Documentation for GeneticAlgorithms.

GeneticAlgorithms.Crossover.SinglePointCrossover

GeneticAlgorithms.Mutation.RealGeneMutation

GeneticAlgorithms.Mutation.SudokuMutation

GeneticAlgorithms.PopulationInitialization.RealUniformInitialization

GeneticAlgorithms.PopulationInitialization.SudokuInitialization

GeneticAlgorithms.Selection.RouletteWheelSelection

GeneticAlgorithms.Fitness.rosenbrock_fitness

GeneticAlgorithms.Fitness.sudoku_fitness

GeneticAlgorithms.Utils.visualize_function_with_contours

GeneticAlgorithms.Utils.visualize_rosenbrock_results

GeneticAlgorithms.optimize

GeneticAlgorithms.optimize — Method

optimize(genetic_algorithm::GeneticAlgorithm)

Runs the genetic algorithm with the specified parameters and returns the best individual found.

GeneticAlgorithms.PopulationInitialization.RealUniformInitialization — Type

RealUniformInitialization(population_size::Int64, chromosome_size::Int64, interval::Tuple{T,T})

Creates a population of population_size including chromosomes of chromosome_size. The chromosome-values are drawn from a uniform distribution over interval. The current implementation supports Float64 and Integer types. The type is determined by the interval.

GeneticAlgorithms.PopulationInitialization.SudokuInitialization — Type

SudokuInitialization(population_size::Int64, initial::Vector{Vector{Int64}})

Creates a population of population_size including chromosomes of 9x9 size. Each gene resembles a column in a Sudoku puzzle. The initial values are taken from the initial grid. initial must be of size 9x9. The remaining values are filled with the missing random values. The initialization ensure that each column contains the values 1-9 exactly once.

GeneticAlgorithms.Selection.RouletteWheelSelection — Method

Performs Roulette Wheel Selection on a population based on fitness scores, returning two selected individuals (parents).

Selection is based on the cumulative probabilities of the fitness scores.

GeneticAlgorithms.Fitness.rosenbrock_fitness — Method

rosenbrock_fitness(values::Chromosome{Float64})::Float64

Calculates 100 divided by the Rosenbrock function value for a given vector.

Arguments

chromosome: A chromosome containing the genes to evaluate.

Returns

100 divided by the Rosenbrock function value as a Float64.

GeneticAlgorithms.Fitness.sudoku_fitness — Method

sudoku_fitness(chromosome::Chromosome{Vector{Int64}})::Float64

Calculates the fitness of a Sudoku puzzle represented by a chromosome. The fitness value is the sum of the number of distinct values in each row and 3x3 subgrid.

GeneticAlgorithms.Mutation.RealGeneMutation — Type

RealGeneMutation(mutation_rate::Float64, mutation_interval::Tuple{T,T})

Defines the mutation of genes of real values (currently only Float64 and Integer types (including Bool)) Mutation is applied for each gene on its own with probability mutation_rate with values from the uniform distribution in the interval mutation_interval.

GeneticAlgorithms.Mutation.SudokuMutation — Type

SudokuMutation(mutation_rate::Float64, initial::Vector{Vector{Int64}})

Mutation method for Sudoku puzzles. The mutation is applied column-wise with probability mutation_rate. initial must be of size 9x9. Ensures that the initial values of initial are not changed and the remaining values are shuffled.

GeneticAlgorithms.Utils.visualize_function_with_contours — Method

Plots a contour of a 2D function over the specified x and y ranges.

Arguments

f::Function: The function to visualize.

x_range: Range of x-values (default: (-2.0, 2.0)).

y_range: Range of y-values (default: (-2.0, 2.0)).

path: Optimization path points (default: []).

Returns

A contour plot of the function with optional highlighted points.

GeneticAlgorithms.Utils.visualize_rosenbrock_results — Function

visualize_rosenbrock_results(best_chromosomes::Vector{Chromosome{Float64}}, save_path::String="result.png")

Generates a visualization of the optimization process for the rosenbrock fitness function with given best chromosomes.

Arguments

best_chromosomes: A sequence of optimization results (chromosomes) for visualizing the optimization path.

Example

julia
visualize_results([Chromosome([0.0, 0.0]), Chromosome([0.5, 0.5]), Chromosome([1.0, 1.0])], "rosenbrock_path.png")

GeneticAlgorithms.Crossover.SinglePointCrossover — Type

SinglePointCrossover

Defines the single point crossover method for two chromosomes.